refactor: rearrangement parameters according to requirements.
