+++ app/src/main/java/com/london/tudee/presentation/components/buttons/TudeeNegativeTextButton.kt
@@ -15,11 +15,11 @@ import com.london.tudee.presentation.design_system.theme.TudeeTheme
 
 @Composable
 fun TudeeNegativeTextButton(
-    onClick: () -> Unit,
     text: String,
     modifier: Modifier = Modifier,
     isDisabled: Boolean = false,
-    isLoading: Boolean = false
+    isLoading: Boolean = false,
+    onClick: () -> Unit
 ) {
     if (isDisabled.not()) {
         Row(
