+++ app/src/main/java/com/london/tudee/presentation/components/priority/PrioritySelector.kt
@@ -11,8 +11,8 @@ import com.london.tudee.presentation.design_system.theme.ThemePreviews
 
 @Composable
 fun PrioritySelector(
-    modifier: Modifier = Modifier,
     selectedPriority: Priority,
+    modifier: Modifier = Modifier,
     onPrioritySelected: (Priority) -> Unit
 ) {
     Row(
