+++ app/src/main/java/com/london/tudee/presentation/design_system/color/TudeeColors.kt
@@ -36,5 +36,6 @@ data class TudeeColors(
     val upperSmallCircleMoon: Color = Color(0xffE9EFFF),
     val onPrimaryCard : Color,
     val onPrimaryStroke : Color,
-    val spotColor : Color = Color(0x0A000000)
+    val spotColor : Color = Color(0x0A000000),
+    val miniBlack : Color = Color(0xFF1F1F1F)
 )
