+++ app/src/main/res/drawable/arrow_right_double.xml
@@ -1,14 +1,7 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="28dp"
-    android:height="28dp"
-    android:viewportWidth="28"
-    android:viewportHeight="28">
-  <path
-      android:pathData="M5.25,7C5.25,6.56 5.498,6.157 5.89,5.959C6.283,5.76 6.755,5.8 7.109,6.061L7.111,6.062C7.165,6.102 7.324,6.221 7.421,6.294C7.616,6.442 7.895,6.654 8.228,6.914C8.894,7.434 9.786,8.15 10.682,8.927C11.571,9.699 12.491,10.556 13.197,11.355C13.549,11.754 13.873,12.165 14.115,12.567C14.338,12.936 14.583,13.444 14.583,14C14.583,14.555 14.338,15.063 14.115,15.432C13.873,15.834 13.549,16.246 13.197,16.644C12.491,17.444 11.571,18.3 10.682,19.072C9.786,19.85 8.894,20.565 8.228,21.085C7.895,21.346 7.616,21.558 7.421,21.706C7.325,21.778 7.168,21.895 7.112,21.937L7.109,21.939C6.754,22.2 6.283,22.24 5.89,22.041C5.498,21.842 5.25,21.44 5.25,21L5.25,7Z"
-      android:fillColor="#ffffff"
-      android:fillAlpha="0.87"/>
-  <path
-      android:pathData="M14.181,8.156C14.087,8.085 13.941,7.976 13.892,7.939C13.373,7.557 13.262,6.827 13.644,6.308C14.026,5.789 14.757,5.679 15.276,6.061L15.277,6.062C15.332,6.103 15.491,6.221 15.587,6.294C15.783,6.442 16.061,6.654 16.395,6.914C17.061,7.434 17.953,8.15 18.848,8.927C19.738,9.699 20.657,10.556 21.364,11.355C21.716,11.754 22.04,12.165 22.282,12.567C22.504,12.936 22.75,13.444 22.75,14C22.75,14.555 22.504,15.063 22.282,15.432C22.04,15.834 21.716,16.246 21.364,16.644C20.657,17.444 19.738,18.3 18.848,19.072C17.953,19.85 17.061,20.565 16.395,21.085C16.063,21.344 15.479,21.785 15.28,21.936L15.276,21.939C14.757,22.321 14.026,22.21 13.644,21.692C13.262,21.173 13.373,20.442 13.892,20.06C14.079,19.919 14.637,19.497 14.959,19.246C15.606,18.741 16.464,18.052 17.319,17.31C18.179,16.563 19.009,15.785 19.615,15.099C19.92,14.755 20.143,14.461 20.283,14.228C20.351,14.116 20.385,14.042 20.402,14C20.385,13.957 20.351,13.883 20.283,13.772C20.143,13.538 19.92,13.245 19.615,12.9C19.009,12.214 18.179,11.436 17.319,10.689C16.464,9.947 15.606,9.258 14.959,8.754C14.637,8.502 14.368,8.297 14.181,8.156Z"
-      android:fillColor="#ffffff"
-      android:fillAlpha="0.87"/>
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:autoMirrored="true" android:height="28dp" android:viewportHeight="28" android:viewportWidth="28" android:width="28dp">
+      
+    <path android:fillAlpha="0.87" android:fillColor="#ffffff" android:pathData="M5.25,7C5.25,6.56 5.498,6.157 5.89,5.959C6.283,5.76 6.755,5.8 7.109,6.061L7.111,6.062C7.165,6.102 7.324,6.221 7.421,6.294C7.616,6.442 7.895,6.654 8.228,6.914C8.894,7.434 9.786,8.15 10.682,8.927C11.571,9.699 12.491,10.556 13.197,11.355C13.549,11.754 13.873,12.165 14.115,12.567C14.338,12.936 14.583,13.444 14.583,14C14.583,14.555 14.338,15.063 14.115,15.432C13.873,15.834 13.549,16.246 13.197,16.644C12.491,17.444 11.571,18.3 10.682,19.072C9.786,19.85 8.894,20.565 8.228,21.085C7.895,21.346 7.616,21.558 7.421,21.706C7.325,21.778 7.168,21.895 7.112,21.937L7.109,21.939C6.754,22.2 6.283,22.24 5.89,22.041C5.498,21.842 5.25,21.44 5.25,21L5.25,7Z"/>
+      
+    <path android:fillAlpha="0.87" android:fillColor="#ffffff" android:pathData="M14.181,8.156C14.087,8.085 13.941,7.976 13.892,7.939C13.373,7.557 13.262,6.827 13.644,6.308C14.026,5.789 14.757,5.679 15.276,6.061L15.277,6.062C15.332,6.103 15.491,6.221 15.587,6.294C15.783,6.442 16.061,6.654 16.395,6.914C17.061,7.434 17.953,8.15 18.848,8.927C19.738,9.699 20.657,10.556 21.364,11.355C21.716,11.754 22.04,12.165 22.282,12.567C22.504,12.936 22.75,13.444 22.75,14C22.75,14.555 22.504,15.063 22.282,15.432C22.04,15.834 21.716,16.246 21.364,16.644C20.657,17.444 19.738,18.3 18.848,19.072C17.953,19.85 17.061,20.565 16.395,21.085C16.063,21.344 15.479,21.785 15.28,21.936L15.276,21.939C14.757,22.321 14.026,22.21 13.644,21.692C13.262,21.173 13.373,20.442 13.892,20.06C14.079,19.919 14.637,19.497 14.959,19.246C15.606,18.741 16.464,18.052 17.319,17.31C18.179,16.563 19.009,15.785 19.615,15.099C19.92,14.755 20.143,14.461 20.283,14.228C20.351,14.116 20.385,14.042 20.402,14C20.385,13.957 20.351,13.883 20.283,13.772C20.143,13.538 19.92,13.245 19.615,12.9C19.009,12.214 18.179,11.436 17.319,10.689C16.464,9.947 15.606,9.258 14.959,8.754C14.637,8.502 14.368,8.297 14.181,8.156Z"/>
+    
 </vector>
